requirements.txt
Flask==2.0.2
psycopg==3.1.18
pymongo==4.0.1
Werkzeug==2.2.2